Pass floats instead of strings into HeatingManagerImpl
Make it the caller’s responsibility to turn strings into floats, as the
caller potentially has better ways to react to non-string inputs (though
the actual caller here isn’t doing anything special).

Ideally, manage()’s $threshold would also be changed to a float, but I
don’t know if we’re allowed to change that or not.